requirements.txt
altgraph==0.17.4
antlr4-python3-runtime==4.13.2
blinker==1.9.0
bottle==0.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.2
cqlib==1.3.4
cycler==0.12.1
dify-client==0.1.10
Flask==3.1.1
fonttools==4.59.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
Nuitka==2.7.12
numpy==2.3.1
openqasm3==1.0.1
ordered-set==4.1.0
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
requests==2.32.4
rustworkx==0.16.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
zstandard==0.23.0
